Build out project template
This commit adds the basic building blocks for a book project of any
type. This includes:

- Icon fonts (ionicons)
- Header nav and associated styles
- Vendor css files
- Book logic and helper methods

The last item represents a stripped down version of the catalogue
extension from Mosaics, represented by a book.rb file and some
associated helper methods. Going forward, this minimal implementation of
book functionality should serve as a template for what can eventually be
released as a Middleman add-on. In light of that I'm attempting to
organize the files in a more modular way (splitting out helper methods,
etc), as well as providing more documentation and writing code with an
eye to reusability.

Currently the next_chapter_path and previous_chapter_path methods are
working as expected, and a few utility methods (like one that gets all
chapters in a book) have been defined as well.