Build out Book extension
Going forward, the Book extension should be defined as a series of files
sharing the Book:: module namespace. Core extension is
Book::BookExtension, chapter module is Book::Chapter, etc. In the future
a Book::Search, Book::PDF, and Book::Epub module can be added in as
well. Some other helper classes like SearchIndexResource (and probably
also templates for the Epub file) will also need to be added. Finally, a
utility class could handle file zipping and archiving.

The goal is to follow classical Object Oriented patterns as much as
possible, but since this code is going to sit on top of the Middleman
codebase there will be times where operating in terms of strict class
inheritance is not possible. For example, as nice as it would be for
Book::Chapter to be a subclass of Book, we need to rely on the
functionality that already exists in the Middleman::Sitemap::Resource
class and can't tamper with the inheritance hierarchy.

The solution is taken from the official Middleman::Blog extension – use
modules and #extend them into specific object instances where necessary.
This is a classic metaprogramming technique in Ruby. Working with
modules will allow for following some conventions in design while not
having to re-invent too many wheels. See the implementation of the
Book::Chapter module in this commit for an example of this technique.