Change nav styles and add link-color mixin
The new link-color() mixin simplifies the process of making links that
transition between two colors (a solid and a slightly transparent
version of the first), for use in places where the background-color is
something other than white. Going forward these mixins can live in the
global stylesheet until there are enough of them to justify being moved
to their own file.

Also, some color changes on the nav bar and sidebar.