Rewrite Book::Chapter as Resource subclass
Based on @tdrenyo's suggestion, this commit re-works how Book::Chapter
objects are represented. Instead of mixing in a module to
already-existing Resource objects, the new approach creates a custom
subclass of Middleman::Sitemap::Resource and replaces any resources
which possess `sort_order` metadata with instances of this class.

This seems to solve the problem caused by LiveReload events under the
previous approach: things would work fine initially but chapters would
"forget" that they had been extended with the additional module on
reload, causing some things to break.

One caveat of rolling your own Resource subclasses - make sure to pass
all the right data to them. In addition to the arguments provided to the
initialize method (store, path, and source file), it is important to
explicitly call the resource#add_metadata method and pass in a
correctly-formatted hash. The easiest way to do this was to just copy
the same hash from the original resource's metadata and pass it in after
initializing the Book::Chapter object.