Implement improved layout and sticky nav sidebar
Lots of style and front-end changes here.

1. Moving away from using Neat for layout grids. The Neat grid does not
play well with fixed-position sticky elements (things break) and it
seems like more trouble than it is worth to debug this. Instead, a
flexbox-based layout seems to be a better approach going forward.

Right now Neat is still being used for breakpoints, but in the future
the entire library could probably be removed.

Similar CSS structure is being maintained, with a .container class that
serves as the display: flex parent (it used to get outer-container
applied).

2. Add in dynamic sidebar nav
Currently all "page" layouts get a chapter nav in the sidebar (each book
chapter appears there in order). Current chapter is styled differently
than others and is non-linked.

To add in a sub-section nav, just set a frontmatter property of
sections: true. Then indicate sections in the Markdown using this
format:
{:.section-heading}

Javascript will fill in the section nav on the fly and animate the
scroll effect.