Add a wrapper div for Hammer.js
Hammer.js and SmoothState were both trying to hook onto the #main
element in order to set up, and this was causing some strange behavior
(next and previous links would not update properly after pages changed,
for example). This commit adds a div immediately below the #main element
called #hammer-js-container, which the Hammer instance is initialized
on. This div is the one that listens to swipe events.